+++ Data/New.lua
@@ -31,6 +31,40 @@ Grants Level 20 Summon Petrification Statue Skill
 5% additional Physical Damage Reduction while Moving
 5% reduced Elemental Damage Taken while stationary
 ]],[[
+Mark of the Elder
+Steel Ring
+Requires Level 80
+Adds (3-4) to (10-14) Physical Damage to Attacks
+Adds 29 to 45 Cold Damage to Attacks
+6% increased maximum Energy Shield
+9% increased maximum Life
+75% increased Attack Damage if your other Ring is a Shaper Item
+Cannot be Stunned by Attacks if your other Ring is an Elder Item
+20% chance to Trigger Level 20 Tentacle Whip on Kill
+Elder Item
+]],[[
+Mark of the Red Covenant
+Tribal Circlet
+Requires Level 26
++49 to maximum Energy Shield
+Minions have 12% increased Movement Speed
+Raging Spirits have 144% increased Damage
+75% reduced Maximum number of Summoned Raging Spirits
+Raging Spirits' Hits Always Ignite
+Raging Spirits Refresh their Duration when they Kill an Ignited Enemy
+]],[[
+Mark of the Shaper
+Opal Ring
+Requires Level 80
+(15-25)% increased Elemental Damage
+Adds 15 to 50 Lightning Damage to Spells
+9% increased maximum Energy Shield
+9% increased maximum Life
+79% increased Spell Damage if your other Ring is an Elder Item
+Cannot be Stunned by Spells if your other Ring is a Shaper Item
+20% chance to Trigger Level 20 Summon Volatile Anomaly on Kill
+Shaper Item
+]],[[
 Might of the Meek
 Crimson Jewel
 Radius: Large
@@ -60,6 +94,17 @@ Adds 37 to 100 Chaos Damage to Attacks
 30% increased Projectile Speed
 5 Maximum Void Charges
 Elder Item
+]],[[
+Winterweave
+Coral Ring
+Requires Level 53
++(20-30) to maximum Life
+Adds 12 to 29 Fire Damage to Attacks
+Adds 12 to 27 Cold Damage to Attacks
++24 to maximum Life
++26% to Cold Resistance
+10% increased Movement Speed while Ignited
+The Effect of Chill on you is reversed
 ]],
 
 -- 3.2 Fated
@@ -87,6 +132,18 @@ Adds (10-20) to (30-50) Cold Damage
 (30-40)% increased Cold Damage with Attack Skills
 Trigger Level 20 Icicle Nova when you Kill a Frozen Enemy
 ]],[[
+Crystal Vault
+Latticed Ringmail
+Requires Level 49
++15 to Strength
++15 to Intelligence
+(140-160)% increased Armour and Energy Shield
++(50-75)% to Cold Resistance
+Cannot be Chilled
+20% of Physical Damage from Hits taken as Cold Damage
+30% of Fire Damage from Hits taken as Cold Damage
+18% increased Effect of Chill
+]],[[
 The Dancing Duo
 Reaver Sword
 Requires Level 65
@@ -110,6 +167,39 @@ Adds 54 to 79 Chaos Damage to Spells
 +67 to maximum Energy Shield
 +50 to Total Mana Cost of Skills
 ]],[[
+Dreadbeak
+Rusted Sword
+Requires Level 61
+18% increased Accuracy Rating
+100% increased Damage when on Low Life
+50% increased Physical Damage
+Adds 94 to 140 Physical Damage
+10% increased Attack Speed
++29 to maximum Life
++2 Life gained for each Enemy hit by Attacks
+You have Onslaught while on Low Life
+]],[[
+Dreadsurge
+Cleaver
+Requires Level 60
+Adds (5-15) to (20-25) Physical Damage
+Adds 227 to 269 Fire Damage
+(7-10)% increased Attack Speed
++(15-25)% to Fire Resistance
+5% increased Movement Speed
+Hits ignore Enemy Monster Fire Resistance while you are Ignited
+]],[[
+Duskblight
+Ironscale Boots
+Requires Level 40
+Socketed Gems are Supported by Level 15 Added Chaos Damage
+141% increased Armour and Evasion
++60 to maximum Life
+20% increased Movement Speed
+50% increased Stun and Block Recovery
+20% reduced Light Radius
++50% to Chaos Resistance during any Flask Effect
+]],[[
 Fox's Fortune
 Wild Leather
 Requiresd level 54
@@ -121,6 +211,27 @@ You gain 1500 Evasion Rating when on Full Life
 10% increased Movement Speed
 Enemies are Unlucky when Damaging you while you are on Full Life
 ]],[[
+Frostferno
+Leather Hood
+Requires Level 60
++3 to Level of Socketed Fire Gems
++3 to Level of Socketed Cold Gems
+Socketed Gems are Supported by Level 30 Cold to Fire
+470% increased Evasion Rating
+60% increased Mana Regeneration Rate
++(20-30)% to Fire and Cold Resistances
+]],[[
+Geofri's Devotion
+Brass Maul
+Requires Level 61
+20% increased Stun Duration on Enemies
+Trigger Level 20 Elemental Warding when you HIt an Enemy while Cursed
+200% increased Physical Damage
+Adds 55 to 77 Physical Damage
+Adds 11 to 23 Cold Damage
+13% increased Stun Duration on Enemies
+Never deal Critical Strikes
+]],[[
 Greedtrap
 Velvet Slippers
 Requires Level 52
@@ -131,8 +242,20 @@ Requires Level 52
 12% increased Quantity of Items found when on Low Life
 20% increased Movement Speed
 ]],[[
+Mirebough
+Gnarled Branch
+Requires Level 32
+18% Chance to Block
+Socketed Gems are Supported by Level 16 Trap
+Socketed Gems are Supported by Level 16 Cluster Trap
+Socketed Gems are Supported by Level 16 Trap and Mine Damage
+48% increased Damage
+20% increased maximum Life
+20% increased maximum Mana
+]],[[
 The Nomad
 Studded Belt
+Requires Level 48
 (20-30)% increased Stun Duration on Enemies
 (25-40)% increased Global Physical Damage
 +(40-50) to Strength
@@ -163,8 +286,21 @@ Cannot be Shocked while Chilled
 30% chance to Chill Attackers for 4 seconds on Block
 35% chance to Shock Attackers for 4 seconds on Block
 ]],[[
+Sunspite
+Clasped Boots
+Requires Level 59
+10% increased Attack Speed
+289% increased Evasion and Energy Shield
+(20-30)% increased Rarity of Items found
++(30-40)% to Fire Resistance
+10% increased Movement Speed
+17% increased Movement Speed while Ignited
++25% chance to be Ignited
+121.9 Life Regenerated per second while Ignited
+]],[[
 The Tactician
 Studded Belt
+Requires Level 48
 (20-30)% increased Stun Duration on Enemies
 (25-40)% increased Global Physical Damage
 +(40-50) to Strength
@@ -172,6 +308,28 @@ Studded Belt
 50% increased Flask Charges gained
 +(20-25)% to all Elemental Resistances while you have at least 200 Strength
 60% increased Critical Strike Chance while you have at least 200 Intelligence
+]],[[
+Timetwist
+Moonstone Ring
+Requires Level 64
++(15-25) to maximum Energy Shield
+(10-15)% increased Attack Speed
+(5-10)% increased Cast Speed
++42 to maximum Energy Shield
+15% increased Mana Regeneration Rate
+3% increased Skill Effect Duration
+Unaffected by Temporal Chains
+]],[[
+Wildwrap
+Strapped Leather
+Requires Level 57
+5% increased Attack Speed
++603 to Evasion Rating
++(30-40) to Cold Resistance
+5% increased Movement Speed
+(20-25)% increased Mana Recovery from Flasks
+15% increased Dexterity
+1% increased Damage per 15 Dexterity
 ]],[[
 Windshriek
 Reinforced Greaves